Refactors pubsub protocol wire, added a shared static lib and merged the byte swap functions.
