+++ bundles/pubsub/pubsub_protocol/pubsub_protocol_lib/CMakeLists.txt
@@ -1,0 +1,22 @@
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
+# 
+#   http://www.apache.org/licenses/LICENSE-2.0
+# 
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied.  See the License for the
+# specific language governing permissions and limitations
+# under the License.
+
+add_library(celix_pubsub_protocol_lib STATIC
+        src/pubsub_wire_protocol_common.c
+)
+target_link_libraries(celix_pubsub_protocol_lib PUBLIC Celix::pubsub_spi)
+target_include_directories(celix_pubsub_protocol_lib PUBLIC include)
